change datatype of Version property in IDomainEvent.cs

+++ Code/Src/Domain/Neo.Domain/Models/StreamEventTypes/IStreamEventTypeRepository.cs
@@ -6,6 +6,6 @@ namespace Neo.Domain.Models.StreamEventTypes;
 public interface IStreamEventTypeRepository : IDomainRepository
 {
     Task<StreamEventType> GetBy(StreamEventTypeId id, CancellationToken cancellationToken);
-    Task<StreamEventType> GetBy(StreamEventTypeId id, int version, CancellationToken cancellationToken);
+    Task<StreamEventType> GetBy(StreamEventTypeId id, long version, CancellationToken cancellationToken);
     Task Add(StreamEventTypeId id, StreamEventType streamEventType, CancellationToken cancellationToken);
 }
